tools/lightmapper/requirements.txt
ipykernel
numpy
PyOpenGL
glfw
matplotlib
pathlib
pillow
tqdm
opencv-python
pyrr